requirements.txt
pip==25.1.1
pydantic==2.11.4
pydantic_core==2.33.2
pydantic-settings==2.4.0
requests==2.32.4
datamodel-code-generator==0.28.5
PyYAML==6.0.1